+++ app/src/main/java/com/example/juancarlos/quicksaapp/NuevoViajeActivity.java
@@ -125,11 +125,11 @@ public class NuevoViajeActivity extends AppCompatActivity {
 
 
         try {
-            int sumapbruto = Integer.parseInt(pbruto);
-            int sumatara = Integer.parseInt(tara);
+            Double sumapbruto = Double.parseDouble(pbruto);
+            Double sumatara = Double.parseDouble(tara);
 
 
-            int neto = sumapbruto - sumatara;
+            Double neto = sumapbruto - sumatara;
             if (neto >= 0) {
 
                 String pneto = String.valueOf(neto);
@@ -163,10 +163,10 @@ public class NuevoViajeActivity extends AppCompatActivity {
         try {
             String pbruto = editTextPbruto.getText().toString();
             String tara = editTextTara.getText().toString();
-            int sumapbruto = Integer.parseInt(pbruto);
-            int sumatara = Integer.parseInt(tara);
+            Double sumapbruto = Double.parseDouble(pbruto);
+            Double sumatara = Double.parseDouble(tara);
 
-            int neto = sumapbruto - sumatara;
+            Double neto = sumapbruto - sumatara;
             if (neto >= 0) {
                 String pneto = String.valueOf(neto);
                 editTextPneto.setText(pneto);
